# backend/docs/swagger.yaml
basePath: /admin-api
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /ai-platform/projects/{id}/evaluation-runs/{runId}/slices:
    post:
      responses: {}
      security:
      - BearerAuth: []
      summary: Save a governed evaluation slice by category, size, scene, device,
        time and confidence
      tags:
      - VisionAI Evaluation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"